fine tuning data type for sparing memory/limitating spill

+++ RecoLocalTracker/ClusterSplitterGPU/plugins/Cluster_test.h
@@ -49,7 +49,7 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE::Splitting {
                   double forceXError_,
                   double forceYError_,
                   float vertexX, float vertexY, float vertexZ, float vertexEta, float vertexPhi,
-                  bool verbose_, bool debugMode, int targetDetId, int targetClusterOf,
+                  bool verbose_, bool debugMode, int targetDetId, uint16_t targetClusterOf,
                   Queue& queue);
 
 }  // namespace ALPAKA_ACCELERATOR_NAMESPACE::Splitting
